.ipynb_checkpoints/requirements-checkpoint.txt
datasets==1.1.3
importlib-metadata==2.0.0
matplotlib==3.3.2
multiprocess==0.70.11.1
nltk==3.5
numpydoc==1.1.0
packaging==20.4
protobuf==3.13.0
py==1.9.0
Pygments==2.7.1
pyparsing==2.4.7
pytest==6.1.0
python-dateutil==2.8.1
scikit-learn==0.23.2
scipy==1.5.2
sentencepiece==0.1.91
tokenizers==0.8.1rc2
Unidecode==1.1.1
word2number==1.1
filelock==3.0.12
spacy==3.2.3
fasttext==0.9.2

